+++ app/client/src/WidgetQueryGenerators/MongoDB.ts
@@ -9,11 +9,11 @@ enum COMMAND_TYPES {
 }
 const ALLOWED_INITAL_VALUE_KEYS = ["aggregate", "smartSubstitution"];
 
-export default class MongoDB extends BaseQueryGenerator {
-  private buildBasicConig(command: COMMAND_TYPES, tableName: string) {
+export default abstract class MongoDB extends BaseQueryGenerator {
+  private static buildBasicConfig(command: COMMAND_TYPES, tableName: string) {
     return { command: { data: command }, collection: { data: tableName } };
   }
-  private buildSelect(combinedConfig: CombinedConfig) {
+  private static buildSelect(combinedConfig: CombinedConfig) {
     const { config, select } = combinedConfig;
 
     if (!select) return;
@@ -41,74 +41,85 @@ export default class MongoDB extends BaseQueryGenerator {
     if (isEmpty(buildCommand)) return;
 
     return {
-      find: buildCommand,
-      ...this.buildBasicConig(COMMAND_TYPES.FIND, config.tableName),
+      actionTitle: "Find_query",
+      actionPayload: {
+        find: buildCommand,
+        ...this.buildBasicConfig(COMMAND_TYPES.FIND, config.tableName),
+      },
     };
   }
-  private buildUpdate(combinedConfig: CombinedConfig) {
+  private static buildUpdate(combinedConfig: CombinedConfig) {
     const { config, insert } = combinedConfig;
 
     if (!insert || !insert.where) return;
 
     return {
-      updateMany: {
-        query: { data: insert.where },
-        update: { data: insert.value },
+      actionTitle: "Update_query",
+      actionPayload: {
+        updateMany: {
+          query: { data: insert.where },
+          update: { data: insert.value },
+        },
+        ...this.buildBasicConfig(COMMAND_TYPES.UPDATE, config.tableName),
       },
-      ...this.buildBasicConig(COMMAND_TYPES.UPDATE, config.tableName),
     };
   }
-  private buildInsert(combinedConfig: CombinedConfig) {
+  private static buildInsert(combinedConfig: CombinedConfig) {
     const { config, create } = combinedConfig;
 
     if (!create || !create.value) return;
 
     return {
-      insert: {
-        documents: { data: create.value },
+      actionTitle: "Insert_query",
+      actionPayload: {
+        insert: {
+          documents: { data: create.value },
+        },
+        ...this.buildBasicConfig(COMMAND_TYPES.INSERT, config.tableName),
       },
-      ...this.buildBasicConig(COMMAND_TYPES.INSERT, config.tableName),
     };
   }
-  removeUnrelatedInitialValues(
-    initialValues: Record<string, any>,
-    commandKey: string,
-  ) {
-    if (isEmpty(initialValues)) return {};
-    return [...ALLOWED_INITAL_VALUE_KEYS, commandKey]
-      .filter((key) => initialValues[key])
-      .reduce((acc, key) => {
-        acc[key] = initialValues[key];
-        return acc;
-      }, {} as Record<string, any>);
-  }
 
-  mergeWithRelatedInitialValues(
+  private static createPayload(
     formDataInitialValues: Record<string, any>,
     commandKey: string,
     builtValues: Record<string, any> | undefined,
   ) {
-    if (!builtValues || isEmpty(builtValues)) return;
+    if (!builtValues || isEmpty(builtValues)) {
+      return;
+    }
     // if no initial values return payload with builtin values
-    if (!formDataInitialValues || isEmpty(formDataInitialValues))
+    if (!formDataInitialValues || isEmpty(formDataInitialValues)) {
       return builtValues;
-    const scubedOutInitalValues = this.removeUnrelatedInitialValues(
-      formDataInitialValues,
-      commandKey,
-    );
+    }
 
-    return { formData: merge(scubedOutInitalValues, builtValues) };
+    const scrubedOutInitalValues = [...ALLOWED_INITAL_VALUE_KEYS, commandKey]
+      .filter((key) => formDataInitialValues[key])
+      .reduce((acc, key) => {
+        acc[key] = formDataInitialValues[key];
+        return acc;
+      }, {} as Record<string, any>);
+
+    const { actionPayload, actionTitle } = builtValues;
+    return {
+      actionPayload: {
+        formData: merge({}, scrubedOutInitalValues, actionPayload),
+      },
+      actionTitle,
+    };
   }
-  build(
+  static build(
     combinedConfig: CombinedConfig,
     pluginInitalValues: { actionConfiguration: any },
   ) {
     const allBuildConfigs = [];
+
     const formDataInitialValues =
       pluginInitalValues?.actionConfiguration?.formData;
+
     if (combinedConfig.select) {
       allBuildConfigs.push(
-        this.mergeWithRelatedInitialValues(
+        this.createPayload(
           formDataInitialValues,
           "find",
           this.buildSelect(combinedConfig),
@@ -117,7 +128,7 @@ export default class MongoDB extends BaseQueryGenerator {
     }
     if (combinedConfig.insert) {
       allBuildConfigs.push(
-        this.mergeWithRelatedInitialValues(
+        this.createPayload(
           formDataInitialValues,
           "updateMany",
           this.buildUpdate(combinedConfig),
@@ -126,7 +137,7 @@ export default class MongoDB extends BaseQueryGenerator {
     }
     if (combinedConfig.create) {
       allBuildConfigs.push(
-        this.mergeWithRelatedInitialValues(
+        this.createPayload(
           formDataInitialValues,
           "insert",
           this.buildInsert(combinedConfig),
